+++ visualization/app/codeCharta/state/store/appSettings/appSettings.actions.ts
@@ -29,7 +29,7 @@ import { defaultExperimentalFeaturesEnabled } from "./enableExperimentalFeatures
 import { defaultLayoutAlgorithm } from "./layoutAlgorithm/layoutAlgorithm.actions"
 import { defaultMaxTreeMapFiles } from "./maxTreeMapFiles/maxTreeMapFiles.actions"
 import { defaultSharpnessMode } from "./sharpnessMode/sharpnessMode.actions"
-import { defaultClipboardEnabled } from "./enableClipboard/clipboardEnabled.actions"
+import { defaultScreenshotToClipboardEnabled } from "./enableClipboard/screenshotToClipboardEnabled.actions"
 
 export enum AppSettingsActions {
 	SET_APP_SETTINGS = "SET_APP_SETTINGS"
@@ -76,7 +76,7 @@ export const defaultAppSettings: AppSettings = {
 	searchPanelMode: defaultSearchPanelMode,
 	sortingOrderAscending: defaultSortingOrderAscending,
 	experimentalFeaturesEnabled: defaultExperimentalFeaturesEnabled,
-	clipboardEnabled: defaultClipboardEnabled,
+	screenshotToClipboardEnabled: defaultScreenshotToClipboardEnabled,
 	layoutAlgorithm: defaultLayoutAlgorithm,
 	maxTreeMapFiles: defaultMaxTreeMapFiles,
 	sharpnessMode: defaultSharpnessMode