+++ visualization/app/codeCharta/state/store/appSettings/enableClipboard/clipboardEnabled.splitter.ts
@@ -1,5 +1,0 @@
-import { setClipboardEnabled } from "./clipboardEnabled.actions"
-
-export function splitClipboardEnabledAction(payload: boolean) {
-	return setClipboardEnabled(payload)
-}