+++ visualization/app/codeCharta/util/customConfigBuilder.ts
@@ -70,7 +70,7 @@ function initializeAppSettings(target: CustomConfig) {
 		showOnlyBuildingsWithEdges: false,
 		sortingOrderAscending: false,
 		experimentalFeaturesEnabled: false,
-		clipboardEnabled: false,
+		screenshotToClipboardEnabled: false,
 		layoutAlgorithm: undefined,
 		maxTreeMapFiles: 0,
 		sharpnessMode: undefined,