+++ visualization/app/codeCharta/util/globalSettingsHelper.ts
@@ -9,7 +9,7 @@ import { setSharpnessMode } from "../state/store/appSettings/sharpnessMode/sharp
 import { setMaxTreeMapFiles } from "../state/store/appSettings/maxTreeMapFiles/maxTreeMapFiles.actions"
 import { setResetCameraIfNewFileIsLoaded } from "../state/store/appSettings/resetCameraIfNewFileIsLoaded/resetCameraIfNewFileIsLoaded.actions"
 import packageJson from "../../../package.json"
-import { setClipboardEnabled } from "../state/store/appSettings/enableClipboard/clipboardEnabled.actions"
+import { setScreenshotToClipboardEnabled } from "../state/store/appSettings/enableClipboard/screenshotToClipboardEnabled.actions"
 
 export class GlobalSettingsHelper {
 	static readonly GLOBALSETTINGS_LOCAL_STORAGE_ELEMENT = "globalSettings"
@@ -41,8 +41,8 @@ export class GlobalSettingsHelper {
 			if (appSettings.experimentalFeaturesEnabled !== globalSettings.experimentalFeaturesEnabled) {
 				storeService.dispatch(setExperimentalFeaturesEnabled(globalSettings.experimentalFeaturesEnabled))
 			}
-			if (appSettings.clipboardEnabled !== globalSettings.clipboardEnabled) {
-				storeService.dispatch(setClipboardEnabled(globalSettings.clipboardEnabled))
+			if (appSettings.screenshotToClipboardEnabled !== globalSettings.screenshotToClipboardEnabled) {
+				storeService.dispatch(setScreenshotToClipboardEnabled(globalSettings.screenshotToClipboardEnabled))
 			}
 			if (appSettings.layoutAlgorithm !== globalSettings.layoutAlgorithm) {
 				storeService.dispatch(setLayoutAlgorithm(globalSettings.layoutAlgorithm))